Extract Cookie Consents Cache Instance
